Move ValidationResult, ValidationError, ValidationSeverity from SPI to common/core for re-usability
