+++ embabel-agent-api/src/main/kotlin/com/embabel/common/core/validation/ValidationSeverity.kt
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.spi.validation
+package com.embabel.common.core.validation
 
-import com.embabel.agent.core.AgentScope
-import com.embabel.common.core.validation.ValidationResult
-
-interface AgentValidationManager {
-    fun validate(agentScope: AgentScope): ValidationResult
-    fun validateWithDetails(agentScope: AgentScope): DetailedValidationResult
+enum class ValidationSeverity {
+    ERROR,
+    CRITICAL,
+    WARNING,
+    INFO
 }
